requirements.txt
# Database
mysql-connector-python>=8.0.0
python-dotenv>=0.19.0

# Caching
redis>=4.0.0

# GUI
tkinter

# Development
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0

# Optional: For enhanced features
pydantic>=1.10.0
python-dateutil>=2.8.0
